[Add]画像と文字の合成部分をテンプレートごとにコンポーネント化した
