+++ app/javascript/components/NewPostPage/NewPostPage.vue
@@ -1,136 +1,98 @@
 <template>
-	<div>
-	  <!-- Templates Slider Start -->
-    <div class="card rounded-3 my-5">
-	    <!-- Title -->
-	    <div class="card-header text-center">
-		    <h3>テンプレートの選択</h3>
-	    </div>
-	    <!-- Body -->
-	    <div class="card-body d-flex flex-column justify-content-center">
-		    <!-- Template Slider-->
-	      <div class="thumb-example">
-          <!-- swiper1 -->
-    	    <swiper class="swiper gallery-top" :options="swiperOptionTop" ref="swiperTop" @slideChange="getActiveSlide">
-	          <swiper-slide v-for="template in templates" >
-		          <img :src="template.img">
-  	        </swiper-slide>
-            <div class="swiper-pagination"  slot="pagination"></div>
-					</swiper>
-        	<!-- swiper2 Thumbs -->
-          <swiper class="swiper gallery-thumbs" :options="swiperOptionThumbs" ref="swiperThumbs">
-	          <swiper-slide v-for="template in templates">
-		          <img :src="template.img">
-	          </swiper-slide>
-          </swiper>
-        </div>
-        <div class="text-end">
-          <input type="button" value="選択する" class="btn btn-warning mt-3" @click="showTemplate">
-        </div>
-    	</div>
-  	</div>
-	  <!-- Templates Slider End -->
-		<!-- Create Image Stard -->
-		<div class="card rounded-3 my-5">
-	    <!-- Title -->
-	    <div class="card-header text-center">
-		    <h3>画像と文字の合成</h3>
-	    </div>
-	    <!-- Body -->
-	    <div class="card-body d-flex flex-column justify-content-center">
-				<div class="form-group">
-					<label for="pet_image">ペットの画像を選択</label>
-	  			<input type="file" class="form-control" ref="input" id="pet_image" name="image" accept="image/*" @change="setImage"/>
-        </div>
-				<div v-show="imgSrc" class="mt-3">
-          <vue-cropper
-						ref="cropper"
-						:src="imgSrc"
-						:auto-crop-area="0.5"
-						:aspect-ratio="1 / 1"
-					/>
-					<div class="text-end">
-            <button class="btn btn-warning mt-3" @click.prevent="drawCroppedImg">トリミング</button>
+	<!-- Container START -->
+	<div class="container">
+		<!-- Row START -->
+		<div class="row justify-content-center vh-100 py-5">
+			<!-- Main content START -->
+			<div class="col-sm-10 col-md-8 col-lg-7">
+	
+				<!-- Templates Slider START -->
+				<div class="card rounded-3 my-5">
+					<!-- Title -->
+					<div class="card-header text-center">
+						<h3>テンプレートの選択</h3>
 					</div>
-		  	</div>
-				<div class="mt-3" v-show="cropImg">
-					<div class="form-group mt-1" >
-						<label for="pet_name">ペットの名前</label>
-						<input type="text" class="form-control" id="pet_name" placeholder="ポチ"></input>
+					<!-- Body -->
+					<div class="card-body d-flex flex-column justify-content-center">
+						<!-- Template Slider-->
+						<div class="thumb-example">
+							<!-- swiper1 -->
+							<swiper class="swiper gallery-top" :options="swiperOptionTop" ref="swiperTop" @slideChange="getActiveSlide">
+								<swiper-slide v-for="template in templates" >
+									<img :src="template.img">
+								</swiper-slide>
+								<div class="swiper-pagination"  slot="pagination"></div>
+							</swiper>
+							<!-- swiper2 Thumbs -->
+							<swiper class="swiper gallery-thumbs" :options="swiperOptionThumbs" ref="swiperThumbs">
+								<swiper-slide v-for="template in templates">
+									<img :src="template.img">
+								</swiper-slide>
+							</swiper>
+						</div>
 						<div class="text-end">
-							<button @click="drawPetName" class="btn btn-warning mt-3">合成</button>
+							<input type="button" value="選択する" class="btn btn-warning mt-3" @click="showTemplate">
 						</div>
 					</div>
-					<div class="form-group mt-1">
-						<label for="crime_name">犯行内容・罪状</label>
-						<input type="text" class="form-control" id="crime_name" placeholder="脱走罪・ブランドバッグ破壊罪・可愛すぎ罪など"></input>
-						<div class="text-end">
-							<button @click="drawCrimeName" class="btn btn-warning mt-3">合成</button>
-						</div> 
+				</div>
+				<!-- Templates Slider END -->
+
+				<!-- Create Image START -->
+				<Template0 v-if="selectedSlide === 0"></Template0>
+				<Template1 v-if="selectedSlide === 1"></Template1>
+				<Template2 v-if="selectedSlide === 2"></Template2>
+        <!-- Create Image END -->
+
+				<!-- New Post START -->
+				<div class="card rounded-3 my-5">
+					<!-- Title -->
+					<div class="card-header text-center">
+						<h3>画像の投稿</h3>
 					</div>
-					<div class="form-group  mt-1">
-						<label for="bounty">懸賞金額</label>
-						<input type="text" class="form-control" id="bounty" placeholder="¥30,000,000"></input>
-						<div class="text-end">
-							<button @click="drawBounty" class="btn btn-warning mt-3">合成</button>
-						</div> 
+					<!-- Body -->
+					<div class="card-body d-flex flex-column justify-content-center">
+						<canvas id="combined_canvas" width="1200" height="630"></canvas>
+						<!-- Form START -->
+						<form action="/posts" accept-charset="UTF-8" method="post">
+							<!-- csrfトークンをhiddenで設定 -->
+							<input type="hidden" id="authenticity_token" name="authenticity_token" value="" autocomplete="off" />
+							<!-- 最終イメージをhiddenで設定 -->
+							<input type="hidden" id="post_image" name="post[image]" value="" />
+							<!-- キャプション -->
+							<div class="form-group">
+								<label for="post_body">キャプション(任意)</label>
+								<textarea class="form-control" name="post[body]" id="post_body"></textarea>
+							</div>
+							<div class="text-end">
+								<button type="submit" class="btn btn-warning mt-3" data-disable-with="投稿する">投稿する</button>
+							</div>
+						</form>
+						<!-- Form END -->
 					</div>
 				</div>
-				<div class="mt-3">
-					<label for="canvas-wrapper">合成イメージ</label>
-					<div class="canvas-wrapper mt-3">
-						<canvas id="image_canvas" width="1200" height="630"></canvas>
-						<canvas id="text_canvas" width="1200" height="630"></canvas>
-					</div>
-					<div class="text-end">
-  				  <button v-show="cropImg" @click="setCompletedImage('#combined_canvas', ['#image_canvas', '#text_canvas'])" class="btn btn-warning mt-3">合成を完了</button>
-					</div>
-			  </div>
-			</div>      
-    </div>
-	  <!-- Create Image End -->
-		<!-- New Post Start -->
-    <div class="card rounded-3 my-5">
-	    <!-- Title -->
-	    <div class="card-header text-center">
-		    <h3>画像の投稿</h3>
-	    </div>
-	    <!-- Body -->
-	    <div class="card-body d-flex flex-column justify-content-center">
-		    <canvas id="combined_canvas" width="1200" height="630"></canvas>
-				<!-- Form Start -->
-				<form action="/posts" accept-charset="UTF-8" method="post">
-					<!-- csrfトークンをhiddenで設定 -->
-					<input type="hidden" id="authenticity_token" name="authenticity_token" value="" autocomplete="off" />
-					<!-- 最終イメージをhiddenで設定 -->
-					<input type="hidden" id="post_image" name="post[image]" value="" />
-					<!-- キャプション -->
-					<div class="form-group">
-						<label for="post_body">キャプション(任意)</label>
-            <textarea class="form-control" name="post[body]" id="post_body"></textarea>
-					</div>
-					<div class="text-end">
-					  <button type="submit" class="btn btn-warning mt-3" data-disable-with="投稿する">投稿する</button>
-				  </div>
-        </form>
-				<!-- Form End -->
-    	</div>
-  	</div>
-	  <!-- Templates Slider End -->
-  </div>
+				<!-- New Post END -->
+			</div>
+			<!-- Main content END -->
+		</div>
+		<!-- Row END -->
+	</div>
+	<!-- Container END -->
 </template>
 
 <script>
   import { Swiper, SwiperSlide } from 'vue-awesome-swiper'
   import 'swiper/css/swiper.css'
-	import VueCropper from 'vue-cropperjs';
-  import 'cropperjs/dist/cropper.css';
+	import Template0 from "./Template0.vue";
+	import Template1 from "./Template1.vue";
+	import Template2 from "./Template2.vue";
 
 	export default {
 		components: {
       Swiper,
       SwiperSlide,
-			VueCropper
+			Template0,
+			Template1,
+			Template2,
     },
     data: function () {
       return {
@@ -155,9 +117,7 @@
 	    		slidesPerView: 3,
 					touchRatio: 0.2,
 	    		slideToClickedSlide: true
-	    	},
-				imgSrc: '',
-        cropImg: '',			
+	    	}		
       }
     },
 		mounted() {
@@ -176,37 +136,6 @@
 			getActiveSlide(){
 				this.selectedSlide = this.$refs.swiperTop.$swiper.realIndex
 			},
-			setImage (e) {
-				const file = e.target.files[0]
-				if (!file.type.includes('image/')) {
-					alert('画像ファイルを選択して下さい')
-					return
-				}
-				if (typeof FileReader === 'function') {
-					const reader = new FileReader();
-          reader.onload = (event) => {
-						this.imgSrc = event.target.result;
-						// rebuild cropperjs with the updated source
-          	this.$refs.cropper.replace(event.target.result);
-          };
-          reader.readAsDataURL(file);
-				} else {
-					alert('Sorry, FileReader API not supported');
-				}
-			},
-			drawCroppedImg() {
-				// トリミングした画像のURLをcropImgに格納
-				this.cropImg = this.$refs.cropper.getCroppedCanvas().toDataURL();
-				// トリミングした画像を選択されたテンプレートの上に合成する
-				const croppedImg = new Image();
-				croppedImg.src = this.cropImg;
-				croppedImg.onload = () =>{
-					const canvas = document.querySelector("#image_canvas");
-					const ctx = canvas.getContext("2d");
-					// 描画の位置は仮設定
-					ctx.drawImage(croppedImg, 70, 120, 500, 500);
-        }
-    	},
       showTemplate() {
 				const selectedTemplate = new Image();
 				selectedTemplate.src = this.templates[this.selectedSlide].img;
@@ -215,64 +144,6 @@
 					const ctx = canvas.getContext("2d");
 					ctx.drawImage(selectedTemplate, 0, 0, canvas.width, canvas.height);
         }
-			},
-			drawPetName(){
-				const canvas = document.querySelector("#text_canvas");
-				const ctx = canvas.getContext('2d');
-				const petName = document.querySelector("#pet_name");
-				//消去の位置は仮設定
-				ctx.clearRect(0, 0, 1200, 280);
-				//スタイルは仮設定
-				ctx.font = '70px serif';
-				ctx.fillStyle = '#404040';
-				//描画の位置は仮設定
-				ctx.fillText(petName.value, 700, 250, 370);
-			},
-			drawCrimeName(){
-				const canvas = document.querySelector("#text_canvas");
-				const ctx = canvas.getContext('2d');
-				const crimeName = document.querySelector("#crime_name");
-				//消去の位置は仮設定
-				ctx.clearRect(0, 300, 1200, 150);
-				//スタイルは仮設定
-				ctx.font = '70px serif';
-				ctx.fillStyle = '#404040';
-				//描画の位置は仮設定
-				ctx.fillText(crimeName.value, 700, 400, 370);
-			},
-			drawBounty(){
-				const canvas = document.querySelector("#text_canvas");
-				const ctx = canvas.getContext('2d');
-				const bounty = document.querySelector("#bounty");
-				//消去の位置は仮設定
-				ctx.clearRect(0, 450, 1200, 150);
-				//スタイルは仮設定
-				ctx.font = '70px serif';
-				ctx.fillStyle = '#404040';
-				//描画の位置は仮設定
-				ctx.fillText(bounty.value, 700, 550, 370);
-			},
-			setCompletedImage:async function(base, assets){
-				// imageとtextの2つのcanvasを合成する
-				const canvas = document.querySelector(base);
-        const ctx = canvas.getContext("2d");
-
-				for(let i=0; i<assets.length; i++){
-					const image1 = await this.getImagefromCanvas(assets[i]);
-					ctx.drawImage(image1, 0, 0, canvas.width, canvas.height);
-				}
-			  // 完成イメージをフォーム内のhiddenに設定
-				const imageUrl = canvas.toDataURL('image/jpeg');
-				document.querySelector("#post_image").value = imageUrl
-			},
-			getImagefromCanvas(id){
-				return new Promise((resolve, reject) => {
-					const image = new Image();
-					const ctx = document.querySelector(id).getContext("2d");
-					image.onload = () => resolve(image);
-					image.onerror = (e) => reject(e);
-					image.src = ctx.canvas.toDataURL();
-				});
 			}
 		}
 	}
@@ -305,19 +176,5 @@
     &.gallery-thumbs .swiper-slide-active {
       opacity: 1;
     }
-  }
-
-	.canvas-wrapper {
-    position: relative;
-		width: 100%;
-		padding-top: 52.5%;
-		height:auto;
-  }
-
-	.canvas-wrapper canvas {
-		position: absolute;
-		width: 100%;
-    top: 0;
-    left: 0;
   }
 </style>
