+++ app/views/posts/new.html.erb
@@ -1,14 +1,2 @@
-<!-- Container START -->
-<div class="container">
-  <div class="row justify-content-center vh-100 py-5">
-    <!-- Main content START -->
-    <div class="col-sm-10 col-md-8 col-lg-7">
-      <!-- Vue START -->
-      <div id="js-new-post-page"></div>
-    </div>
-    <!-- Main content END -->
-  </div>
-  <!-- Row END -->
-</div>
-<!-- Container END -->
-
+<!-- Vue START -->
+<div id="js-new-post-page"></div>
